cyberteck-develop/user-auth/requirements.txt
aniso8601==8.0.0
click==7.1.1
Flask==1.1.2
flask-marshmallow==0.11.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
flask-swagger-ui==3.25.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
pytz==2019.3
six==1.14.0
smmap==3.0.2
SQLAlchemy==1.3.16
Werkzeug==1.0.1
flask_cors==1.10.3
PyJWT==1.7.1
Flask-JWT-Extended==3.24.1
pycryptodome==3.10.1
cryptography==2.9.2
Flask-Cors==1.10.3
